fix(nimble): Fix various issues for security/vulnerability in host

+++ nimble/host/src/ble_hs_hci_priv.h
@@ -88,8 +88,8 @@ int ble_hs_hci_cmd_tx(uint16_t opcode, const void *cmd, uint8_t cmd_len,
 void ble_hs_hci_init(void);
 void ble_hs_hci_deinit(void);
 
-void ble_hs_hci_set_le_supported_feat(uint32_t feat);
-uint32_t ble_hs_hci_get_le_supported_feat(void);
+void ble_hs_hci_set_le_supported_feat(uint64_t feat);
+uint64_t ble_hs_hci_get_le_supported_feat(void);
 void ble_hs_hci_set_hci_version(uint8_t hci_version);
 uint8_t ble_hs_hci_get_hci_version(void);
